Minor fixes

+++ libs/Assets/Sources/common/erm/assets/models/MeshUtils.cpp
@@ -11,7 +11,7 @@ Mesh CreateTriangle(
 	const vec3& c /* = vec3(-1.0f, 1.0f, 0.0f) */
 )
 {
-	constexpr u8 verticesCount = 3;
+	constexpr u32 verticesCount = 3;
 	std::vector<VertexData> vertices(verticesCount);
 	vertices[0].mPositionVertex = a;
 	vertices[0].mNormalVertex = NormalVertex(0.0f, 0.0f, -1.0f);
@@ -20,7 +20,7 @@ Mesh CreateTriangle(
 	vertices[2].mPositionVertex = c;
 	vertices[2].mNormalVertex = NormalVertex(0.0f, 0.0f, -1.0f);
 
-	constexpr u8 indicesCount = 3;
+	constexpr u32 indicesCount = 3;
 	std::vector<IndexData> indices(indicesCount);
 	indices[0] = 0;
 	indices[1] = 1;
@@ -37,7 +37,7 @@ Mesh CreateSquare(
 	const float halfWidth = width * 0.5f;
 	const float halfHeight = height * 0.5f;
 
-	constexpr u8 verticesCount = 4;
+	constexpr u32 verticesCount = 4;
 	std::vector<VertexData> vertices(verticesCount);
 	vertices[0].mPositionVertex = PositionVertex(-halfWidth, -halfHeight, 0.0f);
 	vertices[0].mNormalVertex = NormalVertex(0.0f, 0.0f, -1.0f);
@@ -48,7 +48,7 @@ Mesh CreateSquare(
 	vertices[3].mPositionVertex = PositionVertex(halfWidth, halfHeight, 0.0f);
 	vertices[3].mNormalVertex = NormalVertex(0.0f, 0.0f, -1.0f);
 
-	constexpr u8 indicesCount = 6;
+	constexpr u32 indicesCount = 6;
 	std::vector<IndexData> indices(indicesCount);
 	indices[0] = 0;
 	indices[1] = 1;
@@ -70,10 +70,10 @@ Mesh CreateCube(
 	const float halfY = sizeY * 0.5f;
 	const float halfZ = sizeZ * 0.5f;
 
-	constexpr u8 verticesCount = 6 * 4;
+	constexpr u32 verticesCount = 6 * 4;
 	std::vector<VertexData> vertices(verticesCount);
 
-	u8 vertIndex = 0;
+	u32 vertIndex = 0;
 
 	// Front
 	vertices[vertIndex].mPositionVertex = PositionVertex(-halfX, -halfY, -halfZ);
@@ -160,12 +160,12 @@ Mesh CreateCube(
 	vertices[vertIndex++].mUVVertex = UVVertex(1.0f, 1.0f);
 
 	// Indices
-	constexpr u8 indicesCount = 36;
+	constexpr u32 indicesCount = 36;
 	std::vector<IndexData> indices(indicesCount);
 
-	for (u8 i = 0; i < 6; ++i)
+	for (u32 i = 0; i < 6; ++i)
 	{
-		const u8 offset = i * 4;
+		const u32 offset = i * 4;
 		indices[i * 6] = offset;
 		indices[(i * 6) + 1] = offset + 2;
 		indices[(i * 6) + 2] = offset + 1;
@@ -181,17 +181,17 @@ Mesh CreateSpike(
 	float sizeX /* = 1.0f */,
 	float sizeY /* = 1.0f */,
 	float sizeZ /* = 1.0f */,
-	u8 boneId /* = 0 */
+	u32 boneId /* = 0 */
 )
 {
 	const float halfX = sizeX * 0.5f;
 	const float halfY = sizeY * 0.5f;
 	const float halfZ = sizeZ * 0.5f;
 
-	constexpr u8 verticesCount = 4 + 4 * 3;
+	constexpr u32 verticesCount = 4 + 4 * 3;
 	std::vector<VertexData> vertices(verticesCount);
 
-	u8 vertIndex = 0;
+	u32 vertIndex = 0;
 
 	// Bottom
 	vertices[vertIndex].mPositionVertex = PositionVertex(-halfX, -halfY, halfZ);
@@ -252,7 +252,7 @@ Mesh CreateSpike(
 	vertices[vertIndex++].mNormalVertex = NormalVertex(0.0f);
 
 	// Indices
-	constexpr u8 indicesCount = 18;
+	constexpr u32 indicesCount = 18;
 	std::vector<IndexData> indices(indicesCount);
 
 	indices[0] = 0;
@@ -279,28 +279,28 @@ Mesh CreateSpike(
 
 Mesh CreateSphere(
 	float radius /* = 1.0f */,
-	u8 sectors /* = 10 */,
-	u8 rings /* = 10 */
+	u32 sectors /* = 10 */,
+	u32 rings /* = 10 */
 )
 {
-	const u8 verticesCount = (sectors + 1) * (rings + 1);
+	const u32 verticesCount = (sectors + 1) * (rings + 1);
 	std::vector<VertexData> vertices(verticesCount);
 
-	const u8 indicesCount = sectors * rings * 6 - sectors * 6;
+	const u32 indicesCount = sectors * rings * 6 - sectors * 6;
 	std::vector<IndexData> indices(indicesCount);
 
 	const float sectorStep = static_cast<float>(2.0f * M_PI / sectors);
 	const float stackStep = static_cast<float>(M_PI / rings);
 	float stackAngle = 0.0f;
-	u8 index = 0;
+	u32 index = 0;
 
-	for (u8 i = 0; i <= rings; ++i)
+	for (u32 i = 0; i <= rings; ++i)
 	{
 		stackAngle = static_cast<float>(M_PI * 0.5f - i * stackStep);
 		const float xy = radius * cosf(stackAngle);
 		const float z = radius * sinf(stackAngle);
 
-		for (u8 j = 0; j <= sectors; ++j)
+		for (u32 j = 0; j <= sectors; ++j)
 		{
 			const float sectorAngle = j * sectorStep;
 			const float x = xy * cosf(sectorAngle);
@@ -313,15 +313,15 @@ Mesh CreateSphere(
 		}
 	}
 
-	u8 k1, k2;
+	u32 k1, k2;
 	index = 0;
 
-	for (u8 i = 0; i < rings; ++i)
+	for (u32 i = 0; i < rings; ++i)
 	{
 		k1 = i * (sectors + 1);
 		k2 = k1 + sectors + 1;
 
-		for (u8 j = 0; j < sectors; ++j, ++k1, ++k2)
+		for (u32 j = 0; j < sectors; ++j, ++k1, ++k2)
 		{
 			if (i != 0)
 			{
@@ -343,8 +343,8 @@ Mesh CreateSphere(
 }
 
 Mesh CreateGrid(
-	u8 sizeX /*= 100*/,
-	u8 sizeY /*= 100*/,
+	u32 sizeX /*= 100*/,
+	u32 sizeY /*= 100*/,
 	float width /*= 1.0f*/,
 	float height /*= 1.0f*/
 )
@@ -352,10 +352,10 @@ Mesh CreateGrid(
 	const i8 halfSizeX = sizeX / 2;
 	const i8 halfSizeY = sizeY / 2;
 
-	const u8 verticesCount = (sizeX + 1) * 2 + (sizeY + 1) * 2;
+	const u32 verticesCount = (sizeX + 1) * 2 + (sizeY + 1) * 2;
 	std::vector<VertexData> vertices(verticesCount);
 
-	u8 index = 0;
+	u32 index = 0;
 
 	for (i8 i = 0; i < static_cast<i8>(sizeX + 1); ++i)
 	{
@@ -381,19 +381,19 @@ Mesh CreateGrid(
 		++index;
 	}
 
-	const u8 indicesCount = verticesCount;
+	const u32 indicesCount = verticesCount;
 	std::vector<IndexData> indices(indicesCount);
 
 	index = 0;
 
-	for (u8 i = 0; i < sizeX + 1; ++i)
+	for (u32 i = 0; i < sizeX + 1; ++i)
 	{
 		indices[index * 2] = i;
 		indices[(index * 2) + 1] = i + (sizeX + 1);
 		++index;
 	}
 
-	for (u8 i = 0; i < sizeY + 1; ++i)
+	for (u32 i = 0; i < sizeY + 1; ++i)
 	{
 		const unsigned int start = i + (sizeX + 1) * 2;
 		indices[index * 2] = start;
