+++ libs/Assets/Sources/common/erm/assets/models/MeshUtils.h
@@ -26,16 +26,16 @@ extern Mesh CreateSpike(
 	float sizeX = 1.0f,
 	float sizeY = 1.0f,
 	float sizeZ = 1.0f,
-	u8 boneId = 0);
+	u32 boneId = 0);
 
 extern Mesh CreateSphere(
 	float radius = 1.0f,
-	u8 stackCount = 10,
-	u8 sectorCount = 10);
+	u32 stackCount = 10,
+	u32 sectorCount = 10);
 
 extern Mesh CreateGrid(
-	u8 sizeX = 100,
-	u8 sizeY = 100,
+	u32 sizeX = 100,
+	u32 sizeY = 100,
 	float width = 1.0f,
 	float height = 1.0f);
